fix: remove unnecessary cast in shared_ui.dart
Agent-Logs-Url: https://github.com/ninocss/UntisPlus/sessions/6fc2595d-37a1-4b3c-9fc1-c5ea454f0729

+++ lib/core/shared_ui.dart
@@ -265,10 +265,9 @@ Future<T?> _showUnifiedOptionSheet<T>({
             12 + 16 + 5 + 16 + (subtitle == null ? 58 : 86) + 14;
         final maxListHeight =
             (safeViewportHeight - staticChromeHeight).clamp(
-                  220.0,
-                  safeViewportHeight * 0.82,
-                )
-                as double;
+              220.0,
+              safeViewportHeight * 0.82,
+            );
         final estimatedContentHeight =
             (options.length * 56.0) +
             ((options.length - 1).clamp(0, options.length) * 10.0);
